+++ quil-py/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.8.1-rc.0
+
+### Fixes
+
+- Program equality is sensitive to the order of calibration instructions. (#357)
+
 ## 0.8.0
 
 ### Breaking Changes